+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * expanded the `messageSender` docs on `LlmChatView` to make it clear what it's for and when it's used
 
+* renamed FatXxx to ToolkitXxx e.g. FatColors => ToolkitColors
+
 ## 0.6.3
 
 * fixed [#55](https://github.com/csells/flutter_ai_toolkit/issues/55): TextField overflow error for large inputs